+++ CHANGELOG.md
@@ -315,6 +315,13 @@
 * [X] breeze-core 重构TODO方法。
 * [X] 更新项目文档。
 * [X] 上传到远程仓库。
+
+## 1.1.x
+
+### 1.1.0
+ 
+* [X] 更新版本。更新Kotlin版本到1.3.70。
+* [X] breeze-core 移除集合类型的`randomOrNull`方法，其他一些更新。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。 
 * [ ] breeze-unstable 新的集合类型：`Tree`。
 * [ ] breeze-unstable 新的集合类型：`Table`。
@@ -341,8 +348,8 @@
 
 * [X] 移除对非框架类的第三方库的依赖。
 * [X] 上传到Github。
-* [X] ~~同步项目到Bintray。~~
-* [ ] 发布项目到JCenter或MavenCentral。（直到项目稳定之前不考虑）
+* [X] 同步项目到Bintray。
+* [X] 发布项目到JCenter或MavenCentral。
 * [ ] 检查`awesome-kotlin`中可参考的、有必要参考的项目。
 * [ ] 完善各个模块的`README.md`文档。
 * [ ] 按照功能而非类型更改项目代码的目录结构。